added some new comments and figured out where all the new algorithm parts have to reside
Former-commit-id: f081bc2ae22bc0aa15c0804f1aeaec1b0a0ab01a

+++ tools/clang/lib/CodeGen/CodeGenModule.cpp
@@ -508,7 +508,7 @@ llvm::User* sd_unfold_map_cb(llvm::User* root,
       for (int i = 0; i < children.size(); i++) {
         auto childConst = dyn_cast<llvm::Constant>(children[i]);
         if (!childConst) {
-          unsigned int indices[1] = { i } ;
+          unsigned int indices[1] = { (unsigned int) i } ;
           newStruct = llvm::InsertValueInst::Create(newStruct, children[i], indices, "", arg.insertPos);
         }
       }
